fix timeouts, checking bat level, show bat voltage

+++ components/periodic_task/include/periodic_task.h
@@ -7,6 +7,8 @@ extern "C" {
 #endif
 
 
+#include "stdint.h"
+
 #define FOREVER -1
 
 typedef void(*periodic_func_t)();
@@ -18,10 +20,10 @@ typedef void(*periodic_func_t)();
 void remove_isr_task(periodic_func_t func);
 void remove_task(periodic_func_t func);
 int create_periodic_isr_task(periodic_func_t func,
-                            unsigned delay_ms, 
+                            uint64_t delay_ms, 
                             unsigned count);
 int create_periodic_task(periodic_func_t func,
-                            unsigned delay_sec, 
+                            uint64_t delay_sec, 
                             unsigned count);
 void restart_timer();
 long long get_timer_ms();
